requirements.txt
streamlit>=1.28.0
firebase-admin>=6.0.0
pyrebase4>=4.7.1
google-cloud-firestore>=2.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
protobuf>=4.24.0
requests>=2.31.0
